add totient function
